make it much easier to actually compile this thing
Added new submodules: ring-ffi and cryptography. I'll probably regret
that but for now it gives a way to have python actually build the rust
library and move it to the right place during install and cryptography
means we'll have the vectors and tests to be able to run via tox.

This also adds rust_ext, which gives us some conveniences for building
rust libs automatically via setuptools. No support for develop/editable
though.